backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
alembic>=1.11.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pymysql>=1.1.0
pyodbc>=5.0.0
cx-oracle>=8.3.0

